+++ packages/react-auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/ui-react-auth
 
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies [[`75bf5408c195fa1745bc88f9329c29c66258f247`](https://github.com/aws-amplify/amplify-ui/commit/75bf5408c195fa1745bc88f9329c29c66258f247)]:
+  - @aws-amplify/ui-react-core@3.3.5
+  - @aws-amplify/ui-react@6.9.5
+  - @aws-amplify/ui-react-core-auth@0.0.43
+
 ## 0.0.47
 
 ### Patch Changes